+++ src/routes/auth.routes.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import login  from "../controllers/authController";
+import login  from "../controllers/auth.controller";
 
 const router = express.Router();
